Begin work to render output
